requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
asgiref==3.9.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudscraper==1.2.71
deprecation==2.1.0
dotenv==0.9.9
Flask==3.1.1
frozenlist==1.7.0
gotrue==2.12.2
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
multidict==6.6.0
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
playwright==1.53.0
pluggy==1.6.0
postgrest==1.1.1
prompt_toolkit==3.0.51
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot
pytz==2025.2
realtime==2.5.3
redis==4.6.0
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
telegram==0.0.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
